+++ src/pages/projects/[project_id]/text_messages.tsx
@@ -1,6 +1,8 @@
+import axios, { AxiosRequestConfig } from "axios";
 import { GetServerSideProps, NextPage } from "next";
 
-import { getTextMessages, TextMessagesPage } from "../../../projects/TextMessagesPage";
+import { backendApiUrl } from "../../../../api";
+import { TextMessagesPage } from "../../../projects/TextMessagesPage";
 import { TextMessage } from "../../../projects/type";
 
 type Props = {
@@ -29,4 +31,29 @@ export const getServerSideProps: GetServerSideProps<Props> = async (ctx) => {
       items,
     },
   };
+};
+
+type GetTextMessagesRequest = {
+  limit?: number;
+};
+
+type GetTextMessagesResponse = {
+  count: number;
+  items: TextMessage[];
+};
+
+const getTextMessages = async (project_id: string): Promise<GetTextMessagesResponse> => {
+  const params: GetTextMessagesRequest = {
+    limit: 100, // NOTE: 100という数字は適当
+  };
+
+  const requestConfig: AxiosRequestConfig<GetTextMessagesRequest> = {
+    url: `${backendApiUrl}/api/projects/${project_id}/text_messages/`,
+    method: "GET",
+    params,
+  };
+
+  const { data, status } = await axios.request<GetTextMessagesResponse>(requestConfig);
+  if (status !== 200) throw new Error("failed to get messages");
+  return data;
 };
